chatbot_app/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
strands-agents
mcp
